web28-googleMaps: refactored all code related to setup of map pins into helper fuctions
